% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_fetch_labels.R
\name{api_fetch_labels}
\alias{api_fetch_labels}
\title{api_fetch_labels}
\usage{
api_fetch_labels(
  items,
  language = "en",
  fallback = TRUE,
  APIprefix = "https://www.wikidata.org/w/api.php?action=wbgetentities&"
)
}
\arguments{
\item{items}{character}

\item{language}{ISO 639-1 two-letter language code}

\item{fallback}{boolean: whether to use the Wikidata language fallback chain or not}

\item{APIprefix}{Wikibase API prefix}
}
\value{
data.frame
}
\description{
Fetch item labels from via the Wikibase API.
}
\examples{
api_fetch_labels(items = paste0("Q", 1:100), language = "en", fallback = TRUE, APIprefix = "https://www.wikidata.org/w/api.php?action=wbgetentities&")
}
