% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdfs_copy_to.R
\name{hdfs_copy_to}
\alias{hdfs_copy_to}
\title{hdfs_copy_to}
\usage{
hdfs_copy_to(kerberosUser, localPath, localFilename, hdfsDir)
}
\arguments{
\item{kerberosUser}{character, Kerberos user to impersonate}

\item{localPath}{character, the path to the directory where the localFilename is stored}

\item{localFilename}{character, the local filename}

\item{hdfsDir}{character, the hdfs directory where the files will be placed}
}
\value{

}
\description{
hdfs_copy_to
}
\examples{
hdfs_copy_to(kerberosUser = "", localPath = "", localFilename = "", hdfsDir = "")
}
