% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdfs_mkdir.R
\name{hdfs_mkdir}
\alias{hdfs_mkdir}
\title{hdfs_mkdir}
\usage{
hdfs_mkdir(kerberosUser = "analytics-privatedata", hdfsDir)
}
\arguments{
\item{kerberosUser}{character, the Kerberos user to impersonate}

\item{hdfsDir}{character, hdfsDir: the hdfs directory to mkdir}
}
\value{

}
\description{
hdfs_mkdir
}
\examples{
hdfs_mkdir(kerberosUser = "", hdfsDir = "")
}
