% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdfs_read_from.R
\name{hdfs_read_from}
\alias{hdfs_read_from}
\title{hdfs_read_from}
\usage{
hdfs_read_from(
  kerberosUser,
  localPath,
  localFilenamePrefix,
  hdfsDir,
  hdfsFilenamePrefix,
  fr_header = F
)
}
\arguments{
\item{kerberosUser}{charecter, the Kerberos user to impersonate}

\item{localPath}{character, the path to the directory where the intermediary files should be stored}

\item{localFilenamePrefix}{charcter, how should be the local filenames be prefixed for data.table::fread() to pick them up}

\item{hdfsDir}{character, the hdfs directory where the files are found}

\item{hdfsFilenamePrefix}{charecter,the prefix used for files in the hdfsDir}

\item{fr_header}{boolean, whether or not to use "header = T" in data.table::fread()}
}
\value{

}
\description{
hdfs_read_from
}
\examples{
hdfs_read_from(kerberosUser = "", localPath = "", localFilenamePrefix = "", hdfsDir = "", hdfsFilenamePrefix = "", fr_header = F)
}
