% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdfs_rmdir.R
\name{hdfs_rmdir}
\alias{hdfs_rmdir}
\title{hdfs_rmdir}
\usage{
hdfs_rmdir(kerberosUser = "analytics-privatedata", hdfsDir)
}
\arguments{
\item{kerberosUser}{character, the Kerberos user to impersonate}

\item{hdfsDir}{character, hdfsDir: the hdfs directory to remove}
}
\value{

}
\description{
hdfs_rmdir
}
\examples{
hdfs_rmdir(kerberosUser = "", hdfsDir = "")
}
