% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kerberos_init.R
\name{kerberos_init}
\alias{kerberos_init}
\title{kerberos_init}
\usage{
kerberos_init(kerberosUser = "analytics-privatedata")
}
\arguments{
\item{kerberosUser}{Kerberos user to impersonate}
}
\value{

}
\description{
kerberos_init
}
\examples{
kerberos_init("analytics-privatedata")
}
