% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kerberos_runHiveQL.R
\name{kerberos_runHiveQL}
\alias{kerberos_runHiveQL}
\title{kerberos_runHiveQL}
\usage{
kerberos_runHiveQL(kerberosUser, query, localPath, localFilename)
}
\arguments{
\item{kerberosUser}{kerberosUser}

\item{query}{query}

\item{localPath}{localPath}

\item{localFilename}{localFilename}
}
\value{

}
\description{
Run HiveQL script from the local filesystem on WMF Analytics Client(s)
}
