% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdbs_runSQL.R
\name{rdbs_runSQL}
\alias{rdbs_runSQL}
\title{rdbs_runSQL}
\usage{
rdbs_runSQL(query, database, localPath, localFilename)
}
\arguments{
\item{query}{SQL query}

\item{database}{wiki}

\item{localPath}{path to results (local filesystem)}

\item{localFilename}{result set local filesystem filename}
}
\value{

}
\description{
Run SQL query from WMF Analytics Client(s)
}
\examples{
rdbs_runSQL(query = "show tables;", database = "enwiki", localPath = "/home/goransm/", localFilename = "tables.tsv")
}
