% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_proxy.R
\name{set_proxy}
\alias{set_proxy}
\title{set_proxy}
\usage{
set_proxy(
  http_proxy = "http://webproxy.eqiad.wmnet:8080",
  https_proxy = "http://webproxy.eqiad.wmnet:8080"
)
}
\arguments{
\item{http_proxy}{character: the HTTP proxy}

\item{https_proxy}{character: the HTTPS proxy}
}
\value{

}
\description{
Sets the http and https proxies from the WMF Analytics Clients.
}
\examples{
set_proxy(http_proxy = "http://webproxy.eqiad.wmnet:8080", https_proxy = "http://webproxy.eqiad.wmnet:8080")
}
