% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wdqs_send_query.R
\name{wdqs_send_query}
\alias{wdqs_send_query}
\title{wdqs_send_query}
\usage{
wdqs_send_query(
  query,
  SPARQL_Endpoint = "https://query.wikidata.org/bigdata/namespace/wdq/sparql?query=",
  max_retry = 10
)
}
\arguments{
\item{query}{character, a SPARQL query}

\item{SPARQL_Endpoint}{character, the WDQS endpoint to use}

\item{max_retry}{numeric, number of re-tries to attempt if anything fails}
}
\value{

}
\description{
Function to query WDQS
}
\examples{
wdqs_send_query(query = 'SELECT ?item ?itemLabel WHERE { ?item wdt:P31 wd:Q146. }', SPARQL_Endpoint = "https://query.wikidata.org/bigdata/namespace/wdq/sparql?query=", max_retry = 10)
}
