% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wdqs_superclasses_recurrently.R
\name{wdqs_superclasses_recurrently}
\alias{wdqs_superclasses_recurrently}
\title{wdqs_superclasses_recurrently}
\usage{
wdqs_superclasses_recurrently(
  entity,
  language = "en",
  cleanup = T,
  fetchSubClasses = T,
  fetchCounts = T,
  SPARQL_Endpoint = "https://query.wikidata.org/bigdata/namespace/wdq/sparql?query="
)
}
\arguments{
\item{entity}{character}

\item{language}{ISO 639-1 two-letter language code}

\item{cleanup}{whether to clean up Wikidata URI prefixes or not}

\item{fetchSubClasses}{whether to pick up sub-classes or not}

\item{fetchCounts}{obtain class counts or not}

\item{SPARQL_Endpoint}{the WDQS endpoint to use}
}
\value{
data.frame
}
\description{
Function for recursive search through the Wikidata hierarchy.
Fetch all superclasses (inverse P279 search)
 or parts (inverse P361 search)
up to the hard-coded P279 constraint target: Q35120 (entity)
}
\examples{
wdqs_superclasses_recurrently("Q5", "en", T, T, T, "https://query.wikidata.org/bigdata/namespace/wdq/sparql?query=")
}
